# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AJVC Portfolio Intelligence Dashboard API",
    "description": "\n## 🚀 AJVC Portfolio Intelligence Dashboard - Backend API\n\nA comprehensive **portfolio management system** for venture capital firms to track portfolio companies,\nanalyze project execution, monitor commitments, and gain AI-powered insights from meeting notes.\n\n### Key Features\n\n* **📊 Company Management**: Track portfolio companies with investment details and status\n* **📈 Gantt Layer Analytics**: Pull data from Google Sheets and compute execution KPIs\n* **📝 AI-Powered Meeting Analysis**: Upload DOCX meeting notes for automatic parsing with Google Gemini\n* **✅ Commitment Tracking**: Automatically extract and track actionable commitments from meetings\n* **🔒 Secure Authentication**: JWT-based authentication for all protected endpoints\n\n### Tech Stack\n\n* **FastAPI** for high-performance async API\n* **PostgreSQL** with SQLAlchemy ORM\n* **Google Gemini AI** for intelligent meeting parsing\n* **Google Sheets API** for Gantt data integration\n* **JWT Authentication** for secure access\n\n### Getting Started\n\n1. Authenticate via `/api/v1/auth/login` to get a JWT token\n2. Include the token in the `Authorization: Bearer <token>` header\n3. Create or select a company via `/api/v1/companies`\n4. Upload meeting notes or pull Gantt data for analysis\n\n### API Versioning\n\nCurrent version: **v1.0.0** - All endpoints are prefixed with `/api/v1`\n    ",
...
  "paths": {
...
    "/api/v1/gantt/{company_id}/pull": {
      "post": {
        "tags": [
          "gantt"
        ],
        "summary": "Pull Gantt",
        "operationId": "pull_gantt_api_v1_gantt__company_id__pull_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GanttPullRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GanttPullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gantt/{company_id}/latest": {
      "get": {
        "tags": [
          "gantt"
        ],
        "summary": "Get Latest Snapshot",
        "operationId": "get_latest_snapshot_api_v1_gantt__company_id__latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GanttSnapshotRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gantt/{company_id}/snapshots": {
      "get": {
        "tags": [
          "gantt"
        ],
        "summary": "List Snapshots",
        "operationId": "list_snapshots_api_v1_gantt__company_id__snapshots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GanttSnapshotList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gantt/{company_id}/tasks": {
      "get": {
        "tags": [
          "gantt"
        ],
        "summary": "Get Tasks",
        "operationId": "get_tasks_api_v1_gantt__company_id__tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gantt/{company_id}/velocity-history": {
      "get": {
        "tags": [
          "gantt"
        ],
        "summary": "Get Velocity History",
        "operationId": "get_velocity_history_api_v1_gantt__company_id__velocity_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VelocityHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gantt/portfolio-overview": {
      "get": {
        "tags": [
          "gantt"
        ],
        "summary": "Get Portfolio Overview",
        "description": "Return one row per company with stage counts + KPI scores from their\nlatest GanttSnapshot. Companies with no snapshot are included with\nzero counts and null KPIs.",
        "operationId": "get_portfolio_overview_api_v1_gantt_portfolio_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GanttDiff": {
        "properties": {
          "stage_changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Stage Changes",
            "description": "Tasks that changed stage since last snapshot",
            "default": []
          },
          "new_tasks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "New Tasks",
            "description": "Tasks added since last snapshot",
            "default": []
          },
          "removed_tasks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Removed Tasks",
            "description": "Tasks removed since last snapshot",
            "default": []
          }
        },
        "type": "object",
        "title": "GanttDiff"
      },
      "GanttPullRequest": {
        "properties": {
          "sheets_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheets Url",
            "description": "Google Sheets URL to pull from (overrides company's default)",
            "examples": [
              "https://docs.google.com/spreadsheets/d/1abc...xyz/edit"
            ]
          }
        },
        "type": "object",
        "title": "GanttPullRequest"
      },
      "GanttPullResponse": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id",
            "description": "ID of the created snapshot"
          },
          "upload_date": {
            "type": "string",
            "format": "date",
            "title": "Upload Date",
            "description": "Date of the snapshot"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count",
            "description": "Total number of tasks in the snapshot"
          },
          "shipping_velocity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Velocity",
            "description": "Shipping velocity KPI (0-100)",
            "examples": [
              75.5
            ]
          },
          "execution_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Speed",
            "description": "Execution speed KPI (0-100)",
            "examples": [
              82.3
            ]
          },
          "planning_depth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Depth",
            "description": "Planning depth KPI (0-100)",
            "examples": [
              68.9
            ]
          },
          "planning_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Quality Score",
            "description": "Overall planning quality score (0-100)",
            "examples": [
              71.2
            ]
          },
          "gantt_diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GanttDiff"
              },
              {
                "type": "null"
              }
            ],
            "description": "Diff from previous snapshot"
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "upload_date",
          "task_count"
        ],
        "title": "GanttPullResponse"
      },
      "GanttSnapshotList": {
        "properties": {
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/GanttSnapshotRead"
            },
            "type": "array",
            "title": "Snapshots"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "snapshots",
          "total",
          "page",
          "page_size"
        ],
        "title": "GanttSnapshotList"
      },
      "GanttSnapshotRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "upload_date": {
            "type": "string",
            "format": "date",
            "title": "Upload Date"
          },
          "tasks": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasks"
          },
          "shipping_velocity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Velocity"
          },
          "execution_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Speed"
          },
          "planning_depth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Depth"
          },
          "planning_quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planning Quality Score"
          },
          "task_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Count"
          },
          "gantt_diff": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Gantt Diff"
          },
          "scorecard_history": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scorecard History"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "upload_date",
          "created_at"
        ],
        "title": "GanttSnapshotRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioOverviewResponse": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PortfolioRow"
            },
            "type": "array",
            "title": "Rows",
            "description": "Portfolio company data"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies",
            "description": "Total number of companies"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "total_companies"
        ],
        "title": "PortfolioOverviewResponse"
      },
...
      "TasksResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskObject"
            },
            "type": "array",
            "title": "Tasks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "filters_applied": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Filters Applied"
          }
        },
        "type": "object",
        "required": [
          "tasks",
          "total",
          "filters_applied"
        ],
        "title": "TasksResponse"
      },
...
      "VelocityHistoryResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id",
            "description": "Company ID"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/VelocityPoint"
            },
            "type": "array",
            "title": "History",
            "description": "Historical velocity data points"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "history"
        ],
        "title": "VelocityHistoryResponse"
      },
      "VelocityPoint": {
        "properties": {
          "upload_date": {
            "type": "string",
            "format": "date",
            "title": "Upload Date",
            "description": "Date of the velocity measurement"
          },
          "shipping_velocity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Velocity",
            "description": "Shipping velocity value (0-100)"
          }
        },
        "type": "object",
        "required": [
          "upload_date"
        ],
        "title": "VelocityPoint"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}